+++ CHANGELOG.md
@@ -7,6 +7,12 @@ release date when you use `npm version` (see `README.md`).
 
 ## [Unreleased]
 
+## Added
+
+- An instance when connecting requests information from the backend, which now
+  tells it basic information about the application. This allows us to set the
+  name of the webxdc app in the title of each instance tab.
+
 ## [0.11.0][] - 2022-06-28
 
 ## Added
@@ -228,7 +234,5 @@ release date when you use `npm version` (see `README.md`).
 [0.9.0]: https://github.com/webxdc/webxdc-dev/tree/v0.9.0
 [unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.10.0...HEAD
 [0.10.0]: https://github.com/webxdc/webxdc-dev/tree/v0.10.0
-
-
-[Unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.11.0...HEAD
+[unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.11.0...HEAD
 [0.11.0]: https://github.com/webxdc/webxdc-dev/tree/v0.11.0