+++ CHANGELOG.md
@@ -7,6 +7,8 @@ release date when you use `npm version` (see `README.md`).
 
 ## [Unreleased]
 
+## [0.12.0][] - 2022-06-29
+
 ### Added
 
 - A new mobile tab. This shows the instances in iframes. Console log messages
@@ -245,3 +247,7 @@ release date when you use `npm version` (see `README.md`).
 [0.10.0]: https://github.com/webxdc/webxdc-dev/tree/v0.10.0
 [unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.11.0...HEAD
 [0.11.0]: https://github.com/webxdc/webxdc-dev/tree/v0.11.0
+
+
+[Unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.12.0...HEAD
+[0.12.0]: https://github.com/webxdc/webxdc-dev/tree/v0.12.0